Clean up organization of llvm.rs (#174)
This keeps the code doing the same thing, except in one area: I noticed that floating point literals were only printed to 3 decimal places in the LLVM code, which is bad, so I increased that. In the long term we should use LLVM's hex notation there or use its API to generate code instead of using strings.